# backend/docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: SML Staff Marketplace API
  version: "1.1.0"
  description: |
    Express.js port of the original Java/JAX-RS MarketPlaceWebService.

    All endpoints are under `/service/v1` (included in the server URL below).

    **Auth note**: Failed login returns HTTP 200 with `data: []` — not HTTP 401.

    **Image endpoints**: Return binary `image/png`. `/imagesguid` never returns 404 — returns empty buffer instead.
servers:
  - url: http://localhost:47309/service/v1
    description: Local development (current SML Staff Ubon)
  - url: http://localhost:47300/service/v1
    description: Local development (legacy preview)
...
security:
  - GUIDHeader: []
components:
  securitySchemes:
    GUIDHeader:
      type: apiKey
      in: header
      name: GUID
      description: Session identifier required by most endpoints
  schemas:
    ErrorResponse:
      type: object
      properties:
        ERROR:
          type: string
          example: "database connection error"
...
    CustomerRecord:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        address:
          type: string
        telephone:
          type: string
        tax_id:
          type: string
    EmployeeRecord:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
    CustomerCRMRecord:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        address:
          type: string
        telephone:
          type: string
        logistic_area:
          type: string
        gps:
          type: string
          description: GPS coords stored in ar_customer.website field
        group_main:
          type: string
    CustomerCRMPagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        current_page:
          type: integer
        total_page:
          type: integer
    CustomerCreditData:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        credit_day:
          type: integer
          example: 30
        credit_date:
          type: string
          format: date
          description: Computed as NOW() + credit_day days
    AdvancePayment:
      type: object
      properties:
        doc_no:
          type: string
        doc_date:
          type: string
          format: date
        total_amount:
          type: number
        used:
          type: number
        balance_amount:
          type: number
...
paths:
...
  /getCustomerList:
    get:
      tags: [Customer]
      summary: รายชื่อลูกค้า (max 50)
      description: Dropdown/search — hard limit 50
      parameters:
        - { name: search, in: query, schema: { type: string }, description: "ILIKE ค้นหา code หรือ name_1" }
      responses:
        '200':
          description: รายชื่อลูกค้า
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CustomerRecord' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /getEmployeeList:
    get:
      tags: [Customer]
      summary: รายชื่อพนักงาน (max 50)
      description: เฉพาะ erp_user ที่ name_2='o'
      parameters:
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200':
          description: รายชื่อพนักงาน
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/EmployeeRecord' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /getCustomerCRM:
    get:
      tags: [Customer]
      summary: รายชื่อลูกค้า CRM (paginated)
      description: พร้อม logistic_area, GPS (จาก website field), group_main
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: limit,  in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: รายชื่อลูกค้า + pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CustomerCRMRecord' }
                  pagination: { $ref: '#/components/schemas/CustomerCRMPagination' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /getEmployeeCRM:
    get:
      tags: [Customer]
      summary: รายชื่อพนักงาน (paginated)
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: limit,  in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: รายชื่อพนักงาน
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/EmployeeRecord' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ─── FINANCIAL ───────────────────────────────────────────────────────────────
  /getCustomerCredit:
    get:
      tags: [Financial]
      summary: ข้อมูลเครดิตของลูกค้า
      description: credit_date = NOW() + credit_day วัน (คำนวณใน DB)
      parameters:
        - { name: cust_code, in: query, required: true, schema: { type: string }, example: "C00001" }
      responses:
        '200':
          description: ข้อมูลเครดิต หรือ `{}` ถ้าไม่พบลูกค้า
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    oneOf:
                      - $ref: '#/components/schemas/CustomerCreditData'
                      - type: object
                        description: Empty object when not found
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /getAdvancePayment:
    get:
      tags: [Financial]
      summary: ประวัติมัดจำ/เงินล่วงหน้า
      description: |
        trans_flag IN (40, 9040) — max 20 รายการล่าสุด

        ถ้า last_status=1 → ยอดทั้งหมดแสดงเป็น 0
      parameters:
        - { name: cust_code, in: query, required: true, schema: { type: string }, example: "C00001" }
      responses:
        '200':
          description: รายการมัดจำ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/AdvancePayment' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /getTotalBalance:
    get:
      tags: [Financial]
      summary: ยอดหนี้คงค้างรวม
      description: Invoice (44) + DN (46,93,99,95,101) - CN (48,97,103) - การชำระ ณ date(NOW())
      parameters:
        - { name: cust_code, in: query, required: true, schema: { type: string }, example: "C00001" }
      responses:
        '200':
          description: ยอดหนี้รวม
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:       { type: boolean }
                  total_balance: { type: number, example: 15840.50 }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ─── DOCUMENT ────────────────────────────────────────────────────────────────
...